# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentBroker X",
    "description": "\n**AgentBroker X** is an autonomous agent economy network where AI agents\n**discover, negotiate with, contract, supervise, verify, replace, and pay**\nother AI agents - with no human in the loop.\n\nBuilt for **NANDAHack 2026** (MIT Media Lab + HCLTech).\n\nLifecycle: `discover \u2192 negotiate \u2192 contract \u2192 escrow \u2192 delegate \u2192 verify \u2192\nsettle \u2192 reputation`.\n",
    "contact": {
      "name": "Amaan Khan",
      "url": "https://github.com/amaancoderx"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/reputation/{agent_id}": {
      "get": {
        "tags": [
          "Reputation Engine"
        ],
        "summary": "Get Reputation",
        "operationId": "get_reputation_reputation__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reputation/rate": {
      "post": {
        "tags": [
          "Reputation Engine"
        ],
        "summary": "Rate Agent",
        "operationId": "rate_agent_reputation_rate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/create": {
      "post": {
        "tags": [
          "Negotiation Engine"
        ],
        "summary": "Create Offer",
        "operationId": "create_offer_offers_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/counter": {
      "post": {
        "tags": [
          "Negotiation Engine"
        ],
        "summary": "Counter Offer",
        "operationId": "counter_offer_offers_counter_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferCounter"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/accept": {
      "post": {
        "tags": [
          "Negotiation Engine"
        ],
        "summary": "Accept Offer",
        "operationId": "accept_offer_offers_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/reject": {
      "post": {
        "tags": [
          "Negotiation Engine"
        ],
        "summary": "Reject Offer",
        "operationId": "reject_offer_offers_reject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{offer_id}": {
      "get": {
        "tags": [
          "Negotiation Engine"
        ],
        "summary": "Get Offer",
        "operationId": "get_offer_offers__offer_id__get",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OfferAction": {
        "properties": {
          "offer_id": {
            "type": "string",
            "title": "Offer Id"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id"
          }
        },
        "type": "object",
        "required": [
          "offer_id",
          "actor_id"
        ],
        "title": "OfferAction"
      },
      "OfferCounter": {
        "properties": {
          "offer_id": {
            "type": "string",
            "title": "Offer Id"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "offer_id",
          "actor_id",
          "amount"
        ],
        "title": "OfferCounter"
      },
      "OfferCreate": {
        "properties": {
          "requester_id": {
            "type": "string",
            "title": "Requester Id"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "task_spec": {
            "additionalProperties": true,
            "type": "object",
            "title": "Task Spec"
          },
          "deadline_seconds": {
            "type": "integer",
            "title": "Deadline Seconds",
            "default": 120
          }
        },
        "type": "object",
        "required": [
          "requester_id",
          "provider_id",
          "skill",
          "amount"
        ],
        "title": "OfferCreate"
      },
      "OfferOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "requester_id": {
            "type": "string",
            "title": "Requester Id"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rounds": {
            "type": "integer",
            "title": "Rounds"
          },
          "deadline_seconds": {
            "type": "integer",
            "title": "Deadline Seconds"
          },
          "history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "History"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "requester_id",
          "provider_id",
          "skill",
          "amount",
          "status",
          "rounds",
          "deadline_seconds",
          "history",
          "created_at"
        ],
        "title": "OfferOut"
      },
      "RateRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "rating": {
            "type": "number",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Rating"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "rating"
        ],
        "title": "RateRequest"
      },
...
      "ReputationOut": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "reputation": {
            "type": "number",
            "title": "Reputation"
          },
          "successful_jobs": {
            "type": "integer",
            "title": "Successful Jobs"
          },
          "failed_jobs": {
            "type": "integer",
            "title": "Failed Jobs"
          },
          "total_ratings": {
            "type": "integer",
            "title": "Total Ratings"
          },
          "avg_rating": {
            "type": "number",
            "title": "Avg Rating"
          },
          "trust_tier": {
            "type": "string",
            "title": "Trust Tier"
          },
          "history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "reputation",
          "successful_jobs",
          "failed_jobs",
          "total_ratings",
          "avg_rating",
          "trust_tier",
          "history"
        ],
        "title": "ReputationOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}